feat(ux): consolidate haptics, generate-form targets and type tokens
- remove duplicate useHapticFeedback from mobile-utils; triggerHapticFeedback
  now delegates to lib/haptic (gains Telegram <6.1 version guard, drops any-casts),
  Vibration API fallback preserved
- migrate alert dialog + AddToPlaylistDialog to canonical @/hooks/useHapticFeedback
- delete dead src/hooks/useOffline.ts (0 importers)
- GenerateFormSimple: Palette/AI Boost buttons to 44px targets, clipboard copy
  error handling, px typography to tokens (text-body-lg kills iOS input auto-zoom)
- Sidebar collapsed generation indicator to 44px target
- add UX_AUDIT_2026-07-02.md: journey audit + backlog verification
